extract Price filter view to separate component
